requirements.txt
PyQt6
pillow
google-genai
